spring/spring-core-experiments/spring-resources-examples/src/main/java/org/farrukh/examples/spring/resources/CustomServiceLoader.java
package org.farrukh.examples.spring.resources;

import org.springframework.context.ResourceLoaderAware;
import org.springframework.core.io.Resource;
import org.springframework.core.io.ResourceLoader;

public class CustomServiceLoader implements ResourceLoaderAware {

    private ResourceLoader resourceLoader;

    @Override
    public void setResourceLoader(final ResourceLoader resourceLoader) {
        this.resourceLoader = resourceLoader;
    }

    public Resource getResource(final String location) {
        return resourceLoader.getResource(location);
    }

}
